requirements.txt
streamlit==1.51.0
plotly==6.3.1
pandas==2.3.3
requests==2.32.5
google-generativeai==0.8.5
python-dotenv==1.2.1
